create test_generation_tf_utils.py; Add kwarg check tests
